mktemp: updated args to use clap 3 semantics
After reviewing the GNU mktemp it appears that it requires
the equals for the long form of the TMPDIR arg but not the short.
In order to implement this I've had to split OPT_TMPDIR into
OPT_P and OPT_TMPDIR.

Also added a new dependency "once_cell" with the same version and
uucore. This should make it that the project really isn't gaining any
dependencies overall.